requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.11
streamlit==1.30.0
python-multipart==0.0.6
Pillow==10.0.0
